Do not assume "page" parameters always exist
The Legal Adviser Location Search API this application currently uses
to query postcodes has a pagination component in the response.

The API's expected response on the first page is:

```
{
    "count": 3768,
    "next": "https://staging.laalaa.dsd.io/legal-advisers/?page=2&postcode=BS20BL",
    "previous": null,
    "results": [<snip>
```

The API's expected response on the second page is:

```
{
    "count": 3768,
    "next": "https://staging.laalaa.dsd.io/legal-advisers/?page=3&postcode=BS20BL",
    "previous": "https://staging.laalaa.dsd.io/legal-advisers/?page=1&postcode=BS20BL",
    "results": [<snip>
```

If we upgrade the API application's dependencies then the `previous`
link will change by losing the `page=1` query parameter:

```diff
-"previous": "https://staging.laalaa.dsd.io/legal-advisers/?page=1&postcode=BS20BL",
+"previous": "https://staging.laalaa.dsd.io/legal-advisers/?postcode=BS20BL",
```

The removal of the `page` argument breaks the context filter, as it
makes the assumption that it will always be available:

```
  File ".../cla_public/templates/checker/result/_find-legal-adviser.html", line 97, in top-level template code
    {% set prev_page = data.previous|query_to_dict('page') %}
  File ".../cla_public/apps/base/filters.py", line 35, in query_to_dict
    result = result[prop]
KeyError: 'page'
```

This changeset removes the assumption that "prop"s have to exist.