Add Docker build test to validate expected timezone
Our cla_public containers need to run on Europe/London timezone
due to the code involved around scheduling calls is timezone agnostic.

10:43 in London time is 9:43 in UTC (the current setting), which means
callbacks for the same day would be available from 12:00 (9:43 + 2
hours), instead of the expected 13:00.

Rewriting the code involved is a massive undertaking due to arching into
other shared dependencies. Given these considerations and our nature of
deployment being in UK only, the prudent thing seems to be to ensure the
timezone of our deployments to be set to London.