Check time in a system agnostic way
It is not guaranteed that the `/etc/localtime` file will stay symlinked.
We could also copy the original timezone file and the behaviour would
still be the same, but the test would fail.

This changeset proposes an alternative test for this behaviour, based on
the default format of the `date` command:

```
Fri Feb  2 12:32:42 GMT 2018   # During winter
Tue Aug 28 14:02:43 BST 2018   # During summer
```

Since the command displays the abbreviated name of the timezone, we can
match on those abbreviations.

Note:
- `GMT` stands for Greenwich Mean Time
- `BST` stands for British Summer Time