Configure backend and laalaa URLs consistently
Currently, configuring the `cla_backend` and `laa-legal-adviser-api`
URLs is not straightforward.

Backend can be configured through `BACKEND_API` or
`BACKEND_BASE_URI` or `CLA_BACKEND_PORT`, depending on the environment.

Legal adviser API can be configured through `LAALAA_API_HOST` or
`LAALAA_PORT`, depending on the environment.

This changeset consolidates so that all environments can be configured
via:

- `BACKEND_BASE_URI` environment variable for the `cla_backend` base
  URL, including scheme and port.
- `LAALAA_API_HOST` environment variable for the `laa-legal-adviser-api`
  base URL, including scheme and port.

The default values on the test environment are always on `localhost`.