Skip authenticated steps on forked pull requests
The `$CIRCLE_PR_NUMBER` environment variable is only populated on forked
pull requests. (https://circleci.com/docs/2.0/env-vars/#setting-an-environment-variable-in-a-project)

Using this behaviour we can split the control flow of the Docker build
step to build and test the Docker image, but do not attempt to login and
push.

Initially, I wanted to separate the jobs to (`docker-build` and
`docker-push`) but then I realised making the image built in the first
job available for the second job would not be trivial, so I lazied out.