Set and enforce max length on other_adaptation field
- Adds a max length validator to the field, which is ignored if the
is_other_adaptation field is false
- Removes the `optional` validator from the `adaptations`
ValidatedFormField, as that was short-circuiting validation on the form
- Removes the value of other_adaptation from the form and from the
session if is_other_adaptation is false

As the value of other_adaptation is (rightly) not validated if
is_other_adaptation is false, we can end up passing very long values to
the back end still (if the user fills in the field and then unticks the
box to hide and ignore that field), and it's the use of these in the
session which appears to cause the session to break, leading to the
server error. Therefore, we now remove the value if is_other_adaptation
is false (it was already not passed to the API in this case).