Add badges with links to relevant pages
Our new hires in another team in the service area have highlighted that
it is hard to find something related to an application unless we know
where to look.

One of the ideas to deal with this difficulty is to make the application
Readmes the starting point for everything.

In the above spirit, the new badges introduced will link directly to:

- Where to find the application runbook.
- Where to find the runtime exceptions.
- Where to find the runtime logs.
- Where to find the feature tests.
- Where to find the continuous integration system, with a representation
  of the latest production build.