Send welsh language code to google analytics on welsh pages
A local variable is used to make it more readable as this piece of code is using django templating as well as javascript. 

The code checks if the locale language in the cookies is set to welsh. If it is, it will set the language on the google analytics tracker to welsh. If it is not, it will use the language set as default in the browser.