Match language field formatting used for Welsh and set ENG as a default
The default language code sent to Google Analytics is 'en-gb', which uses a dash instead of an underscore. This code will match the language field set for the Welsh language and make it of the same standard.

Also, by default Google Analytics takes the preferred language set in the browser settings if no language is manually set. Line 188 makes sure that if the locale is not Welsh (cy), set the language field to 'en-gb' and not whatever language is set by the browser.