Refactor code and set GA language field from locale set in cookies
Previously, the code was manually setting the language fields for Welsh and English. The new code changes will take whatever locale is set in the cookies, which ensures consistency.

It also sets a new fallback if no locale is currently set in cookies to en_GB instead of en, also ensuring consistency and making sure the same language code is used for English.