Remove pound sign prefixed as the first character
The pound sign is represented differently depending on where it's coming from ie. from tests or from user input.

The pound sign is already decoded by UTF-8 when coming from the front end (user input side) and represented as '\xa3'. Whereas from tests it is represented as the pound sign symbol. Therefore, the pound sign symbol and \xa3 code point is used to catch all appearances of the pound sign.