Specify exact location of test folder and print test coverage
Using test as an option means the file must contain the word test in it. These new changes mean it doesn't matter the name of the file as long as it is in the correct test folder.

Also the --coverage option will print out the test coverage results in the terminal and create new coverage directory